prepare mis2-2021

+++ cpp/Point3D.h
@@ -10,12 +10,12 @@ protected:
 public:	
 	unsigned hashCode() {
 		// random number from random.org
-		unsigned long long z[] = {0x2058cc50L, 0xcb19137eL, 0x2cb6b6fdL};
-		long zz = 0xbea0107e5067d19dL;
+		uint64_t z[] = {0x2058cc50L, 0xcb19137eL, 0x2cb6b6fdL};
+		uint64_t zz = 0xbea0107e5067d19dL;
 
-		long h0 = ods::hashCode(x0);
-		long h1 = ods::hashCode(x1);
-		long h2 = ods::hashCode(x2);
+		uint64_t h0 = ods::hashCode(x0);
+		uint64_t h1 = ods::hashCode(x1);
+		uint64_t h2 = ods::hashCode(x2);
 		return (unsigned)(((z[0]*h0 + z[1]*h1 + z[2]*h2)*zz) >> 32);
 	}
 };
